Enlarge media cache size to 256MB
Android by default only seems to have a cache size of 20MB which leads to a lot of wasted bandwidth when using ReactMap daily. This aims to use Cronet library to intercept and cache media requests so that we can maintain control over the cache size. This is only supported on Android 14+ or 12-13 with the latest Google Play system update.